1) Several optimizations in wInit to progressively shortlist candidate vectors using distance from polygon. 2)The distance is found from a point that is a convex combination of current set of AE, not a conical combination as was done before

+++ AENMF.cpp
@@ -29,9 +29,9 @@ int main(int argc, char *argv[]) {
 		cerr<<"AENMF N R D inputFile wFile hFile\n";
 		exit(1);
 	}
-	UINT N = strtol(argv[1], NULL, 10);
-	UINT R = strtol(argv[2], NULL, 10);
-	UINT D = strtol(argv[3], NULL, 10);
+	UINT N = (UINT)strtol(argv[1], NULL, 10);
+	UINT R = (UINT)strtol(argv[2], NULL, 10);
+	UINT D = (UINT)strtol(argv[3], NULL, 10);
 	strcpy(inpFname, argv[4]);
 	strcpy(wFname, argv[5]);
 	strcpy(hFname, argv[6]);
